requirements.txt
Flask==3.0.0
flask-cors==4.0.0
Jinja2==3.1.2
Werkzeug==3.0.1
MarkupSafe==2.1.3
Markdown==3.5.1
python-frontmatter==1.0.0
PyYAML==6.0.1